requirements.txt
langgraph>=0.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
